+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## Unreleased
 
+* Add advertised WCS/WFS urls in output Qgis project
+    - Fix https://github.com/3liz/py-qgis-wps/issues/28
 * Fix Sample config (https://github.com/3liz/py-qgis-wps/issues/32)
 * Fix Qgis 3.30 supports
 * Merge pull request #37 from ptitjano/wip/jfelder/complexinput-supportedformats